Hopefully finally shuts ASAN up

+++ retroarch.c
@@ -13124,7 +13124,7 @@ static void input_menu_keys_pressed(input_bits_t *p_new_state,
          {
             for (port = 0; port < port_max; port++)
             {
-               if (binds[port][i].valid && ret[port] & (UINT32_C(1) << i))
+               if (binds[port][i].valid && ret[port] & (UINT64_C(1) << i))
                {
                   bit_pressed = true;
                   break;
@@ -13294,7 +13294,7 @@ static void input_keys_pressed(input_bits_t *p_new_state)
       for (i = 0; i < RARCH_FIRST_META_KEY; i++)
       {
          bool bit_pressed = !input_driver_block_libretro_input 
-            && binds[i].valid && (ret & (UINT32_C(1) <<  i));
+            && binds[i].valid && (ret & (UINT64_C(1) <<  i));
          if (bit_pressed || input_keys_pressed_other_sources(i, p_new_state))
          {
             BIT256_SET_PTR(p_new_state, i);
